HUBBLE-271 wrap contents of search context into own component to provide integration of external search services
